+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
-
+- Add new `LinkBox` JavaScript component.
 ## [v3.1.0] - 2016-12-08
 
 Features: